+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### HEAD
+* Create `wp-cli.local.yml` file with `wp @vagrant` alias for every site ([#1043](https://github.com/roots/trellis/pull/1043))
 * Add Vagrant `ssh-config` to `~/.ssh/config` on `vagrant up` ([#1042](https://github.com/roots/trellis/pull/1042))
 * [BREAKING] Add Ubuntu 18.04 support and default to it ([#992](https://github.com/roots/trellis/pull/992))
 * Python 3 support ([#1031](https://github.com/roots/trellis/pull/1031))